pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "celery>=5.6.2",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "lunarcalendar>=0.0.9",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]
